Fix for  #172
The Windows API specs indicate that this should be
a reference to a long value, which is a 32-bit signed
for both Win32 and Win64. This causes problems for an empty array,
then the upper bound is -1 because long(-1) is not
the same as int64(-1). The casting then incorrectly
returns 4294967295 instead of -1 thus failing subsequent
TotalElements() calculation result returning 4294967296 instead of 0.

When using SafeArrayConversion.ToStringArray(),
for example, this results "out of memory" panic.

Signed-off-by: Artur Troian <troian.ap@gmail.com>

Signed-off-by: Artur Troian <troian.ap@gmail.com>