Fix VT_BOOL value conversion
boolVal is only 16-bits wide in the C structure, but we are reading it
here as a 64-bit integer. The upper bits are not necessarily
initialized. This can lead to "true" being returned incorrectly when the
actual value is "false".

Fix the issse by masking off the upper bits before the comparison.