error.go, idispatch_windows.go: Do a better job of propagating error information from EXCEPINFO.
In go-ole, `OleError` may wrap another error, but it does not surface that
information when returning the error string.

This patch adds that error information to the `OleError` string, and also adds
an `Unwrap` method.

I also cleaned up the instantiation of this error a bit, as it is redundant
to use the inner error's description as the outer error's description now that
we're writing out all the things to the message string.

Signed-off-by: Aaron Klotz <aaron@tailscale.com>