Fix usage of safeArrayCreate (#263)
SafeArrayCreate cast pointer of return value (uintptr) to Pointer of
SafeArray.
Like the other create functions the return-value can directly cast to
Pointer of SafeArray
`
safearray = (*SafeArray)(unsafe.Pointer(&sa))
`

The use of error return values is also misleading. Error is not nil if
the creation of the array has been executed successfully.